requirements.txt
Django==4.2
psycopg2-binary==2.9.9
bcrypt==4.1.2
PyJWT==2.8.0